requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
astroid==2.3.3            # via pylint, pylint-celery
attrs==19.3.0             # via pytest
bleach==3.1.5             # via readme-renderer
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
click-log==0.3.2          # via edx-lint
click==7.1.2              # via click-log, edx-lint
coverage==5.2             # via -r requirements/test.in, pytest-cov
django-dynamic-fixture==3.1.0  # via -r requirements/test.in
docutils==0.16            # via readme-renderer
edx-lint==1.5.0           # via -r requirements/test.in
idna==2.10                # via requests
importlib-metadata==1.7.0  # via pluggy, pytest
isort==4.3.21             # via pylint
lazy-object-proxy==1.4.3  # via astroid
mccabe==0.6.1             # via pylint
mock==3.0.5               # via -c requirements/constraints.txt, -r requirements/test.in
more-itertools==8.4.0     # via pytest
packaging==20.4           # via bleach, pytest
pathlib2==2.3.5           # via pytest
pep8==1.7.1               # via -r requirements/test.in
pkginfo==1.5.0.1          # via twine
pluggy==0.13.1            # via pytest
py==1.9.0                 # via pytest
pygments==2.6.1           # via readme-renderer
pylint-celery==0.3        # via edx-lint
pylint-django==2.0.11     # via edx-lint
pylint-plugin-utils==0.6  # via pylint-celery, pylint-django
pylint==2.4.4             # via edx-lint, pylint-celery, pylint-django, pylint-plugin-utils
pyparsing==2.4.7          # via packaging
pytest-cov==2.10.0        # via -r requirements/test.in
pytest-django==3.9.0      # via -r requirements/test.in
pytest==5.4.3             # via pytest-cov, pytest-django
pytz==2020.1              # via -r requirements/base.txt, django
readme-renderer==26.0     # via twine
requests-toolbelt==0.9.1  # via twine
requests==2.24.0          # via requests-toolbelt, twine
six==1.15.0               # via astroid, bleach, django-dynamic-fixture, edx-lint, mock, packaging, pathlib2, readme-renderer
sqlparse==0.3.1           # via -r requirements/base.txt, django
tqdm==4.47.0              # via twine
twine==1.15.0             # via -r requirements/test.in
typed-ast==1.4.1          # via astroid
urllib3==1.25.9           # via requests
wcwidth==0.2.5            # via pytest
webencodings==0.5.1       # via bleach
wrapt==1.11.2             # via astroid
zipp==1.2.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
